同步

+++ mesh-agent/src/main/java/com/alibaba/dubbo/performance/demo/agent/registry/Endpoint.java
@@ -1,13 +1,13 @@
 package com.alibaba.dubbo.performance.demo.agent.registry;
 
 import java.util.List;
-import java.util.concurrent.atomic.AtomicLong;
+import java.util.concurrent.atomic.AtomicInteger;
 
 public class Endpoint {
     private final String host;
     private final List<Integer> ports;
 
-    public AtomicLong reqNum = new AtomicLong(0);
+    public AtomicInteger reqNum = new AtomicInteger(0);
 
     public Endpoint(String host, List<Integer> ports) {
         this.host = host;
